+++ test/conformance/kernel/urKernelSetArgPointer.cpp
@@ -142,7 +142,7 @@ TEST_P(urKernelSetArgPointerNegativeTest, InvalidNullHandleKernel) {
 }
 
 TEST_P(urKernelSetArgPointerNegativeTest, InvalidKernelArgumentIndex) {
-    size_t num_kernel_args = 0;
+    uint32_t num_kernel_args = 0;
     ASSERT_SUCCESS(urKernelGetInfo(kernel, UR_KERNEL_INFO_NUM_ARGS,
                                    sizeof(num_kernel_args), &num_kernel_args,
                                    nullptr));
